+++ CHANGELOG.md
@@ -40,3 +40,4 @@
 - [feat:upgrade spring cloud 2023 version.](https://github.com/Tencent/spring-cloud-tencent/pull/1451)
 - [feat:support concurrency rate limit.](https://github.com/Tencent/spring-cloud-tencent/pull/1455)
 - [feat:support auth.](https://github.com/Tencent/spring-cloud-tencent/pull/1479)
+- [feat:upgrade trace plugin.](https://github.com/Tencent/spring-cloud-tencent/pull/1480)